Refactor Day09 Part 1
Move into lib.rs so it can be reused by part 2.
